requirements.txt
# run 'pipreqs' from the root directory to generate the requirements.txt file 

beautifulsoup4==4.11.1
dnspython==1.16.0
email-validator==1.1.2
Flask==2.0.3
Flask_Moment==0.11.0
flask_mongoengine==0.9.5
Flask_WTF==1.0.0
Flask_Login==0.5.0
Jinja2==3.0.3
google_api_python_client==2.50.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
# Only needed for running on Google Cloud Run. Does not need to be installed locally.
gunicorn==20.0.0
# Blows up on google Cloud sometimes
#matplotlib==3.3.4
mongoengine==0.20.0
numpy==1.24.3
oauthlib==3.2.0
pandas>=1.5.3
protobuf==4.21.1
pymongo==3.10.1
requests==2.22.0
twilio==6.44.0
Werkzeug==2.0.3
WTForms==2.3.3
WTForms_Components==0.10.4